% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.score.R
\name{bin.score-deprecated}
\alias{bin.score-deprecated}
\title{A Binning Function}
\usage{
bin.score(x, rt.col = "RT", accuracy.col = "Accuracy",
  condition.col = "Condition", baseline.condition = "congruent",
  type = "mean", id = "Subject")
}
\arguments{
\item{x}{a dataframe with trial level data. Needs to have RT and Accuracy DVs}

\item{rt.col}{Column name in dataframe that contains the reaction time data}

\item{accuracy.col}{Column name in dataframe that contains the accuracy data}

\item{condition.col}{Column name in dataframe that contains the trial condition type}

\item{type}{How should Bin trials be aggregated, "sum" or "mean" (Default: "mean")}

\item{baseline.condition}{The values that specify the baseline condition}

\item{id}{Column name in dataframe that contains the subject identifiers}
}
\description{
Calculates Bin scores using the method suggested in [insert citation here]
}
\seealso{
\code{\link{englelab-deprecated}}
}
\keyword{internal}
