% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.score.R
\name{bin.score}
\alias{bin.score}
\title{A Binning Function}
\usage{
bin.score(x, RT.label = "RT", Accuracy.label = "Accuracy",
  Condition.label = "Condition", Subject.label = "Subject",
  baseline.condition = "congruent")
}
\arguments{
\item{x}{a dataframe with trial level data. Needs to have RT and Accuracy DVs}

\item{RT.label}{Column name in dataframe that contains the reaction time data}

\item{Accuracy.label}{Column name in dataframe that contains the accuracy data}

\item{Condition.label}{Column name in dataframe that contains the trial condition type}

\item{Subject.label}{Column name in dataframe that contains the subject identifiers}

\item{baseline.condition}{The values that specify the baseline condition}
}
\description{
Calculates Bin scores using the method suggested in [insert citation here]
}
\examples{
bin.score(data, RT.label = "RT", Accuracy.label = "Accuracy", Condition.label = "TrialType", Subject.label = "Subject")
}
\keyword{bin}
