% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.R
\name{center}
\alias{center}
\title{Center Variables}
\usage{
center(x, variables = c(), standardize = FALSE, drop = FALSE, suffix = NULL)
}
\arguments{
\item{x}{dataframe}

\item{variables}{c() of columns to standardize}

\item{standardize}{Logical. Do you want to calculate
zscores? Default is FALSE}

\item{drop}{Drop original non-centered variables. Default is FALSE}

\item{suffix}{Suffix to add at the end of the column names.
Default is NULL and will add a _c for non-standardized and
_z for standardized z-scores.}
}
\description{
Center variables around the mean. To create z-scores,
specify standardize = TRUE
A new column will be created with the centered values.
}
