% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite.R
\name{composite}
\alias{composite}
\title{Create a Composite Factor}
\usage{
composite(
  x,
  name = NULL,
  variables,
  type = "mean",
  standardize = TRUE,
  missing_allowed = NULL
)
}
\arguments{
\item{x}{data frame}

\item{name}{Name of the new composite variable}

\item{variables}{c() of columns to average together to create the composite}

\item{type}{What type of composite should be calculated?
options are "mean", "sum", "fa" (factor analysis using psych::fa), or
"pca" (principal component analysis using psych::pca). Default is "mean"}

\item{standardize}{Logical. Do you want to calculate the composite based
on standardize (z-score) variables? (Default = TRUE)}

\item{missing_allowed}{Criteria for how many variables can having missing
values and still calculate a composite}
}
\description{
Create a composite factor from a list of variables. A composite factor can be
formed using a mean, sum, factor analysis, or principal components.
}
