% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_project}
\alias{create_project}
\title{Setup directory and templates for a new research study}
\usage{
create_project(wmc_scripts = TRUE, ac_scripts = TRUE, path = ".")
}
\arguments{
\item{wmc_scripts}{Logical. Download WMC task template files?}

\item{ac_scripts}{Logical. Download all attention task template files?}

\item{path}{String. Home directory file path}
}
\description{
This function can be used to automatically setup your study directory by
creating folders and template scripts
}
