% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cronbach.alpha.R
\name{cronbach.alpha-deprecated}
\alias{cronbach.alpha-deprecated}
\title{Calculate Cronbach's Alpha}
\usage{
cronbach.alpha(x, trial.col = "Trial", value = NULL, id = "Subject")
}
\arguments{
\item{x}{a dataframe with trial level data}

\item{trial.col}{The column name that identifies trials}

\item{value}{The column name that identifies the values to be used}

\item{id}{The column name that identifies the Subject IDs.}
}
\description{
Takes a trial-level data frame and calculates cronbach's alpha on the specified value
}
\seealso{
\code{\link{englelab-deprecated}}
}
\keyword{internal}
