% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_distance.R
\name{edit_distance}
\alias{edit_distance}
\title{Calculates EditDistance.unit and EditDistance.load scores}
\usage{
edit_distance(x, target = "MemoryTargets", recall = "Recalled")
}
\arguments{
\item{x}{dataframe}

\item{target}{the column name containing the sequence of target memory items}

\item{recall}{the column name containing the sequence of recalled items}
}
\description{
The Damerau-Levenshtein distance is implemented in the stringdist package:
van der Loo, M. P. J. (2014). The stringdist package for approximate string
matching. R Journal, 6(1), 111-122. doi:10.32614/RJ-2014-011
}
