% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_join.R
\name{files_join}
\alias{files_join}
\title{Merge Multiple Files}
\usage{
files_join(
  path = "",
  pattern = "",
  delim = ",",
  id = "Subject",
  output_file = ""
)
}
\arguments{
\item{path}{Folder location of files to be merged}

\item{pattern}{Pattern string to identify files to be merged}

\item{delim}{Delimiter used in files. Passed onto `readr::read_delim()`}

\item{id}{Subject ID variable name. Passed onto `plyr::join_all(by = id)`}

\item{output_file}{File name and path to be saved to}
}
\description{
Merges multiple files located in the same directory into a single data frame
}
