% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_script.R
\name{get_script}
\alias{get_script}
\title{Download template R scripts}
\usage{
get_script(type = "all", to = "R Scripts", overwrite = FALSE,
  wmc = FALSE, symspan = FALSE, ospan = FALSE, rotspan = FALSE,
  gf = FALSE, rapm = FALSE, numberseries = FALSE,
  lettersets = FALSE, ac = FALSE, antisaccade = FALSE,
  stroop = FALSE, flanker = FALSE, stroopDL = FALSE,
  flankerDL = FALSE, va4 = FALSE, sact = FALSE, path = "./")
}
\arguments{
\item{type}{String. Do you want to download "raw" or "score" task template files?}

\item{to}{a directory where to download R scripts}

\item{wmc}{Logical. Do you want to download template scripts for the wmc taks?}

\item{gf}{Logical. Do you want to download template scripts for the gf taks?}

\item{ac}{Logical. Do you want to download template scripts for all the attention taks?}

\item{antisaccade}{Logical. Do you want to download a template file for the antisaccade task?}

\item{stroop}{Logical. Do you want to download a template file for the Stroop task?}

\item{flanker}{Logical. Do you want to download a template file for the flanker task?}

\item{stroopDL}{Logical. Do you want to download a template file for the StroopDL task?}

\item{flankerDL}{Logical. Do you want to download a template file for the FlankerDL task?}

\item{va4}{Logical. Do you want to download a template file for the Visual Arrays 4 task?}

\item{sact}{Logical. Do you want to download a template file for the SACT task?}

\item{path}{String. Home directory file path}
}
\description{
This function will download R script templates for tasks commonly used in the EngleLab
}
