% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.lettersets.R
\name{raw.lettersets}
\alias{raw.lettersets}
\title{A Task Scoring Function}
\usage{
raw.lettersets(x)
}
\arguments{
\item{x}{a .txt exported E-Merge (or single E-Data) file}
}
\description{
Creates a raw data file of the LetterSets task from the E-Merge file
}
\examples{
raw.lettersets(data)
}
\keyword{raw}
