% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.numberseries.R
\name{raw.numberseries}
\alias{raw.numberseries}
\title{A Task Scoring Function}
\usage{
raw.numberseries(x)
}
\arguments{
\item{x}{a .txt exported E-Merge (or single E-Data) file}
}
\description{
Creates a raw data file of the NumberSeries task from the E-Merge file
}
\examples{
raw.numberseries(data)
}
\keyword{raw}
