% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.rapm.R
\name{raw.rapm}
\alias{raw.rapm}
\title{A Task Scoring Function}
\usage{
raw.rapm(x)
}
\arguments{
\item{x}{a .txt exported E-Merge (or single E-Data) file}
}
\description{
Creates a raw data file of the RAPM task from the E-Merge file
}
\examples{
raw.rapm(data)
}
\keyword{raw}
