% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antisaccade.R
\name{raw_antisaccade}
\alias{raw_antisaccade}
\title{Raw Tidy Data for Antisaccade}
\usage{
raw_antisaccade(x, include_col = c())
}
\arguments{
\item{x}{dataframe}

\item{include_col}{c(): list of additional columns to include}
}
\description{
Converts a messy e-prime data file into a tidy raw data file that is
easy to work with.
}
