% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flankerDL.R
\name{raw_flankerDL}
\alias{raw_flankerDL}
\title{Raw Tidy Data for FlankerDL}
\usage{
raw_flankerDL(x, include_col = c())
}
\arguments{
\item{x}{dataframe}

\item{include_col}{c(): list of additional columns to include}
}
\description{
Converts the messy e-prime data file into a tidy raw data file that is
easy to work with.
}
