% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf.R
\name{raw_lettersets}
\alias{raw_lettersets}
\title{Creates a "tidy" raw dataframe for the LetterSets task}
\usage{
raw_lettersets(x)
}
\arguments{
\item{x}{dataframe (an imported .emrge file)}
}
\description{
Creates a "tidy" raw dataframe for the LetterSets task
}
