% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf.R
\name{raw_numberseries}
\alias{raw_numberseries}
\title{Creates a "tidy" raw dataframe for the NumberSeries task}
\usage{
raw_numberseries(x)
}
\arguments{
\item{x}{dataframe (an imported .emrge file)}
}
\description{
Creates a "tidy" raw dataframe for the NumberSeries task
}
