% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf.R
\name{raw_rapm}
\alias{raw_rapm}
\title{Creates a "tidy" raw dataframe for the RAPM task}
\usage{
raw_rapm(x)
}
\arguments{
\item{x}{dataframe (an imported .emrge file)}
}
\description{
Creates a "tidy" raw dataframe for the RAPM task
}
