% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotspan.R
\name{raw_rotspan}
\alias{raw_rotspan}
\title{Creates a "tidy" raw dataframe for the Rotation Span task}
\usage{
raw_rotspan(x, blocks = "", taskVersion = "new", keep_col = c())
}
\arguments{
\item{x}{dataframe}

\item{blocks}{number of blocks administered. From 1-3}

\item{taskVersion}{old or new version. Old version means the
Procedure[Block] variable has a different label. (Default = "new")}

\item{keep_col}{List of column names to keep in raw dataframe}
}
\description{
Creates a "tidy" raw dataframe for the Rotation Span task
}
