% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sact.R
\name{raw_sact}
\alias{raw_sact}
\title{Raw Tidy Data for SACT}
\usage{
raw_sact(x, include_col = c())
}
\arguments{
\item{x}{dataframe}

\item{include_col}{c(): list of additional columns to include}
}
\description{
Converts the messy e-prime data file into a tidy raw data file that is
easy to work with.
}
