% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squared.R
\name{raw_squared}
\alias{raw_squared}
\title{Raw Tidy Data for Three-Minute Squared Tasks}
\usage{
raw_squared(x, include_col = c())
}
\arguments{
\item{x}{dataframe}

\item{include_col}{c(): list of additional columns to include}
}
\description{
Converts a messy e-prime data file into a tidy raw data file that is
easy to work with.
}
