% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stroopDL.R
\name{raw_stroopDL}
\alias{raw_stroopDL}
\title{Creates a "tidy" raw dataframe for the StroopDL task}
\usage{
raw_stroopDL(x)
}
\arguments{
\item{x}{dataframe (an imported .emrge file)}
}
\description{
Creates a "tidy" raw dataframe for the StroopDL task
}
