% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualarrays.R
\name{raw_visualarrays}
\alias{raw_visualarrays}
\title{Creates a "tidy" raw dataframe for the Visual Arrays task}
\usage{
raw_visualarrays(x)
}
\arguments{
\item{x}{dataframe (an imported .emrge file)}
}
\description{
Creates a "tidy" raw dataframe for the Visual Arrays task
}
