% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_latent.R
\name{remove_latent}
\alias{remove_latent}
\title{Subject Missing Data Removal}
\usage{
remove_latent(x, factor_list = list(), missing_allowed = 1, id = "Subject")
}
\arguments{
\item{x}{dataframe}

\item{factor_list}{list of factors and tasks.
e.g., list(WMC = c("OSpan", "RotSpan", SymSpan"))}

\item{missing_allowed}{Proportion of tasks allowed to be missing}

\item{id}{Subject ID variable}
}
\description{
Remove subjects with too much missing data on a given latent construct
}
