% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_problematic.R
\name{remove_problematic}
\alias{remove_problematic}
\title{Remove Problematic Subjects}
\usage{
remove_problematic(x, filter = "", log_file = NULL)
}
\arguments{
\item{x}{dataframe}

\item{filter}{an expression passed on to dplyr:filter()}

\item{log_file}{the file path and name for where to save a log file to}
}
\description{
This function will remove problematic subjects based on a filter criteria
and save them to a log data file.
}
