% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.lettersets.R
\name{score.lettersets}
\alias{score.lettersets}
\title{A Task Scoring Function}
\usage{
score.lettersets(x)
}
\arguments{
\item{x}{a .txt exported E-Merge (or single E-Data) file}
}
\description{
Scores the Letter Sets task taking a .txt exported E-Merge (or single E-Data) file as input
}
\examples{
score.lettersets(data)
}
\keyword{score}
