% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.numberseries.R
\name{score.numberseries-deprecated}
\alias{score.numberseries-deprecated}
\title{A Task Scoring Function}
\usage{
score.numberseries(x)
}
\arguments{
\item{x}{a .txt exported E-Merge (or single E-Data) file}
}
\description{
Scores the Number Series task taking a .txt exported E-Merge (or single E-Data) file as input
}
\seealso{
\code{\link{englelab-deprecated}}
}
\keyword{internal}
