% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.numberseries.R
\name{score.numberseries}
\alias{score.numberseries}
\title{A Task Scoring Function}
\usage{
score.numberseries(x)
}
\arguments{
\item{x}{a .txt exported E-Merge (or single E-Data) file}
}
\description{
Scores the Number Series task taking a .txt exported E-Merge (or single E-Data) file as input
}
\examples{
score.numberseries(data)
}
\keyword{score}
