% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.ospan.R
\name{score.ospan-deprecated}
\alias{score.ospan-deprecated}
\title{A Task Scoring Function}
\usage{
score.ospan(x, blocks = "")
}
\arguments{
\item{x}{a .txt exported E-Merge (or single E-Data) file}

\item{blocks}{number of blocks administered. From 1-3}
}
\description{
Scores the Operation Span (OSPAN) task taking a .txt exported E-Merge (or single E-Data) file as input
}
\seealso{
\code{\link{englelab-deprecated}}
}
\keyword{internal}
