% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.rapm.R
\name{score.rapm-deprecated}
\alias{score.rapm-deprecated}
\title{A Task Scoring Function}
\usage{
score.rapm(x)
}
\arguments{
\item{x}{a .txt exported E-Merge (or single E-Data) file}
}
\description{
Scores the Raven's Advanced Progressive Matrices (RAPM) task taking a .txt exported E-Merge (or single E-Data) file as input
}
\seealso{
\code{\link{englelab-deprecated}}
}
\keyword{internal}
