% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_adaptivethreshold.R
\name{score_adaptivethreshold}
\alias{score_adaptivethreshold}
\title{Calculates a score for adaptive threshold procedures}
\usage{
score_adaptivethreshold(
  x,
  taskname = "",
  adaptive_variable = "",
  scoring_method = "Last 4 Reversals"
)
}
\arguments{
\item{x}{dataframe}

\item{taskname}{string to add as a prefix to scored columns}

\item{adaptive_variable}{the column name that represents the variable that
is changing across the adaptive procedure (e.g., stimulus duration).}

\item{scoring_method}{the scoring method to use. Default: Last 4 Reversals.
Options are:
Last n Reversals (substitute n for a digit, e.g., Last 4 Reversals. This
represents the median of the last n reversals);
Last Trial (the value of the adaptive variable on the last trial);
All Reversals (median of all reversals);
Overall Median (median across all trials)}
}
\description{
Calculates a score for adaptive threshold procedures
}
