% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antisaccade.R
\name{score_antisaccade}
\alias{score_antisaccade}
\title{Calculate Antisaccacde accuracy scores from a messy raw dataframe}
\usage{
score_antisaccade(x)
}
\arguments{
\item{x}{dataframe (an imported .emrge file)}
}
\description{
This function skips the 'raw_antisaccade()' step and therefore
    is not advised. However, some researchers may find
    it easier to just skip right to 'score_antisaccade()'
}
