% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flankerDL.R
\name{score_flankerDL}
\alias{score_flankerDL}
\title{Calculate FlankerDL threshold scores from a messy raw dataframe}
\usage{
score_flankerDL(x)
}
\arguments{
\item{x}{dataframe (an imported .emrge file)}
}
\description{
This function skips the 'raw_flankerDL()' step and therefore
    is not advised. However, some researchers may find
    it easier to just skip right to 'score_flankerDL()'
}
