% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf.R
\name{score_lettersets}
\alias{score_lettersets}
\title{Calculate LetterSets scores from a messy raw dataframe}
\usage{
score_lettersets(x)
}
\arguments{
\item{x}{dataframe (an imported .emrge file)}
}
\description{
This function skips the 'raw_lettersets()' step and therefore
    is not advised. However, some researchers may find
    it easier to just skip right to 'score_lettersets()'
}
