% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf.R
\name{score_numberseries}
\alias{score_numberseries}
\title{A Task Scoring Function}
\usage{
score_numberseries(x)
}
\arguments{
\item{x}{a .txt exported E-Merge (or single E-Data) file}
}
\description{
Scores the Number Series task taking a .txt exported E-Merge (or single E-Data) file as input
}
\examples{
score_numberseries(data)
}
\keyword{score}
