% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospan.R
\name{score_ospan}
\alias{score_ospan}
\title{Calculate Operation Span Scores}
\usage{
score_ospan(x)
}
\arguments{
\item{x}{dataframe}
}
\description{
Calculate various span scores from the output of `raw_ospan()`
}
