% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf.R
\name{score_rapm}
\alias{score_rapm}
\title{Calculate RAPM scores from a messy raw dataframe}
\usage{
score_rapm(x)
}
\arguments{
\item{x}{dataframe (an imported .emrge file)}
}
\description{
This function skips the 'raw_rapm()' step and therefore
    is not advised. However, some researchers may find
    it easier to just skip right to 'score_rapm()'
}
