% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readspan.R
\name{score_readspan}
\alias{score_readspan}
\title{Calculate Reading Span Scores}
\usage{
score_readspan(x)
}
\arguments{
\item{x}{dataframe}
}
\description{
Calculate various span scores from the output of `raw_readspan()`
}
