% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotspan.R
\name{score_rotspan}
\alias{score_rotspan}
\title{Calculate Rotation Span scores from a messy raw dataframe}
\usage{
score_rotspan(x, blocks = NULL)
}
\arguments{
\item{x}{dataframe}

\item{blocks}{depricated. No need to indicate number of blocks.
Use group_by(Subject, Block) instead}
}
\description{
Calculate Rotation Span scores from a messy raw dataframe
}
