% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sact.R
\name{score_sact}
\alias{score_sact}
\title{Calculate SACT accuracy scores from a messy raw dataframe}
\usage{
score_sact(x)
}
\arguments{
\item{x}{dataframe (an imported .emrge file)}
}
\description{
This function skips the 'raw_sact()' step and therefore
    is not advised. However, some researchers may find
    it easier to just skip right to 'score_sact()'
}
