% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squared.R
\name{score_squared}
\alias{score_squared}
\title{Score Three-Minute Squared Tasks}
\usage{
score_squared(x)
}
\arguments{
\item{x}{dataframe}
}
\description{
Calculates task scores for the Squared Tasks from the output of raw_squared().
}
