% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stroopDL.R
\name{score_stroopDL}
\alias{score_stroopDL}
\title{Calculate StroopDL Scores}
\usage{
score_stroopDL(x, scoring_method = "Last 4 Reversals", rt_cutoff = 0)
}
\arguments{
\item{x}{dataframe}

\item{scoring_method}{List of scoring methods to use, e.g.
c("Last 4 Reversals", "Last 8 Reversals").
Passed onto score_adaptivethreshold().
The scoring method to use. Default: Last 4 Reversals.
Options are:
Last n Reversals (substitute n for a digit, e.g., Last 4 Reversals. This
represents the median of the last n reversals);
Last Trial (the value of the adaptive variable on the last trial);
All Reversals (median of all reversals);
Overall Median (median across all trials)}

\item{rt_cutoff}{Optional. Criterion for cutoff of too short of reaction
times. Used when calculating overall RT performance.}
}
\description{
Calculate StroopDL scores from the output of `raw_stroopDL()`
}
