% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualarrays.R
\name{score_visualarrays}
\alias{score_visualarrays}
\title{Calculate Visual Array k Scores}
\usage{
score_visualarrays(
  x,
  id_col = "Subject",
  taskname = "VAorient_S",
  unit_weight = FALSE
)
}
\arguments{
\item{x}{dataframe}

\item{id_col}{Subject id column name (required)}

\item{taskname}{string to add as a prefix to columns. Useful if your task
includes multiple conditions (excluding set size)}

\item{unit_weight}{logical. If TRUE, the weight of each trial is 1. If FALSE,
the weight of each trial is based on the set size}
}
\description{
Calculate Visual Array k scores from the output of `raw_visualarrays()`
}
