% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{splithalf}
\alias{splithalf}
\title{Calculate Cronbach's Alpha}
\usage{
splithalf(x, trial.col = "Trial", value = "", aggregate = NULL,
  id = "Subject")
}
\arguments{
\item{x}{a dataframe with trial level data}

\item{trial.col}{The column name that identifies trials}

\item{value}{The column name that identifies the values to be used}

\item{aggregate}{How should the value column be aggregated across even and odd trials? (ex. mean or sum)}

\item{id}{The column name that identifies the Subject IDs.}
}
\description{
Takes a trial-level data frame and calculates cronbach's alpha on the specified value
}
\examples{

}
\keyword{cronbach}
