% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_score.R
\name{bin_score}
\alias{bin_score}
\title{A Binning Function}
\usage{
bin_score(
  x,
  rt.col = "RT",
  accuracy.col = "Accuracy",
  condition.col = "Condition",
  baseline.condition = "congruent",
  type = "mean",
  id = "Subject"
)
}
\arguments{
\item{x}{a dataframe with trial level data. Needs to have RT and Accuracy DVs}

\item{rt.col}{Column name in dataframe that contains the reaction time data}

\item{accuracy.col}{Column name in dataframe that contains the accuracy data}

\item{condition.col}{Column name in dataframe that contains the trial condition type}

\item{baseline.condition}{The values that specify the baseline condition}

\item{type}{How should Bin trials be aggregated, "sum" or "mean" (Default: "mean")}

\item{id}{Column name in dataframe that contains the subject identifiers}
}
\description{
Calculates Bin scores using the method suggested in [insert citation here]
}
\examples{
bin_score(data, RT.label = "RT", Accuracy.label = "Accuracy", Condition.label = "TrialType", Subject.label = "Subject")
}
\keyword{bin}
