% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.score.R, R/cronbach.alpha.R,
%   R/englelab-deprecated.r, R/raw.lettersets.R, R/raw.numberseries.R,
%   R/raw.ospan.R, R/raw.rapm.R, R/raw.rotspan.R, R/raw.symspan.R,
%   R/score.lettersets.R, R/score.numberseries.R, R/score.ospan.R,
%   R/score.rapm.R, R/score.rotspan.R
\name{englelab-deprecated}
\alias{englelab-deprecated}
\alias{bin.score}
\alias{cronbach.alpha}
\alias{raw.lettersets}
\alias{raw.numberseries}
\alias{raw.ospan}
\alias{raw.rapm}
\alias{raw.rotspan}
\alias{raw.symspan}
\alias{score.lettersets}
\alias{score.numberseries}
\alias{score.ospan}
\alias{score.rapm}
\alias{score.rotspan}
\title{Deprecated functions in package \pkg{englelab}.}
\usage{
bin.score(x, rt.col = "RT", accuracy.col = "Accuracy",
  condition.col = "Condition", baseline.condition = "congruent",
  type = "mean", id = "Subject")

cronbach.alpha(x, trial.col = "Trial", value = NULL, id = "Subject")

raw.lettersets(x)

raw.numberseries(x)

raw.ospan(x, blocks = "", taskVersion = "new")

raw.rapm(x)

raw.rotspan(x, blocks = "", taskVersion = "new")

raw.symspan(x, blocks = "", taskVersion = "new")

score.lettersets(x)

score.numberseries(x)

score.ospan(x, blocks = "")

score.rapm(x)

score.rotspan(x, blocks = "")
}
\description{
The functions listed below are deprecated and will be defunct in
  the near future. When possible, alternative functions with similar
  functionality are also mentioned. Help pages for deprecated functions are
  available at \code{help("<function>-deprecated")}.
}
\section{\code{bin.score}}{

For \code{bin.score}, use \code{\link{bin_score}}.
}

\section{\code{cronbach.alpha}}{

For \code{cronbach.alpha}, use \code{\link{cronbach_alpha}}.
}

\section{\code{raw.lettersets}}{

For \code{raw.lettersets}, use \code{\link{raw_lettersets}}.
}

\section{\code{raw.numberseries}}{

For \code{raw.numberseries}, use \code{\link{raw_numberseries}}.
}

\section{\code{raw.ospan}}{

For \code{raw.ospan}, use \code{\link{raw_ospan}}.
}

\section{\code{raw.rapm}}{

For \code{raw.rapm}, use \code{\link{raw_rapm}}.
}

\section{\code{raw.rotspan}}{

For \code{raw.rotspan}, use \code{\link{raw_rotspan}}.
}

\section{\code{raw.symspan}}{

For \code{raw.symspan}, use \code{\link{raw_symspan}}.
}

\section{\code{score.lettersets}}{

For \code{score.lettersets}, use \code{\link{score_lettersets}}.
}

\section{\code{score.numberseries}}{

For \code{score.numberseries}, use \code{\link{score_numberseries}}.
}

\section{\code{score.ospan}}{

For \code{score.ospan}, use \code{\link{score_ospan}}.
}

\section{\code{score.rapm}}{

For \code{score.rapm}, use \code{\link{score_rapm}}.
}

\section{\code{score.rotspan}}{

For \code{score.rotspan}, use \code{\link{score_rotspan}}.
}

\examples{
bin.score(data, RT.label = "RT", Accuracy.label = "Accuracy", Condition.label = "TrialType", Subject.label = "Subject")

raw.lettersets(data)
raw.numberseries(data)
raw.ospan(data, blocks = 2)
raw.rapm(data)
raw.rotspan(x, blocks = 2)
raw.symspan(data, blocks = 2)
score.lettersets(data)
score.numberseries(data)
score.ospan(data, blocks = 2)
score.rapm(data)
score.rotspan(x, blocks = 2)
}
\keyword{bin}
\keyword{cronbach}
\keyword{internal}
\keyword{raw}
\keyword{score}
