% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_script.R
\name{get_script}
\alias{get_script}
\title{Download template R scripts}
\usage{
get_script(
  type = "raw",
  to = "R Scripts",
  overwrite = FALSE,
  wmc = FALSE,
  symspan = FALSE,
  ospan = FALSE,
  rotspan = FALSE,
  gf = FALSE,
  rapm = FALSE,
  numberseries = FALSE,
  lettersets = FALSE,
  ac = FALSE,
  antisaccade = FALSE,
  stroop = FALSE,
  flanker = FALSE,
  stroopDL = FALSE,
  flankerDL = FALSE,
  va = FALSE,
  sact = FALSE,
  path = "./"
)
}
\arguments{
\item{type}{String. Do you want to download "raw" or "score" task template files?}

\item{to}{a directory where to download R scripts}

\item{wmc}{Logical. Download scripts for the wmc taks?}

\item{symspan}{Logical. Download script for the symmetry span task?}

\item{ospan}{Logical. Download script for the operation span task?}

\item{rotspan}{Logical. Download script for the rotation span task?}

\item{gf}{Logical. Dowload scripts for the gf taks?}

\item{rapm}{Logical. Download script for ravens advanced progressive matrices?}

\item{numberseries}{Logical. Download script for number series?}

\item{lettersets}{Logical. Download script for letter sets?}

\item{ac}{Logical. Download scripts for all the attention taks?}

\item{antisaccade}{Logical. Download script for the antisaccade task?}

\item{stroop}{Logical. Download script for the Stroop task?}

\item{flanker}{Logical. Download script for the flanker task?}

\item{stroopDL}{Logical. Download script for the StroopDL task?}

\item{flankerDL}{Logical. Download script for the FlankerDL task?}

\item{va}{Logical. Download script for the Visual Arrays task?}

\item{sact}{Logical. Download script for the SACT task?}

\item{path}{String. Home directory file path}
}
\description{
This function will download R script templates for tasks commonly used in the EngleLab
}
