% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antisaccade.R
\name{raw_antisaccade}
\alias{raw_antisaccade}
\title{Creates a "tidy" raw dataframe for the Antisaccade task}
\usage{
raw_antisaccade(x)
}
\arguments{
\item{x}{dataframe (an imported .emrge file)}
}
\description{
Creates a "tidy" raw dataframe for the Antisaccade task
}
