% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antisaccade.R
\name{raw_antisaccade}
\alias{raw_antisaccade}
\title{Creates a "tidy" raw dataframe for the Antisaccade task}
\usage{
raw_antisaccade(x, version = "new")
}
\arguments{
\item{x}{dataframe (an imported .emrge file)}

\item{version}{is this a "new" or "old" version of the task?}
}
\description{
Creates a "tidy" raw dataframe for the Antisaccade task
}
