% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flankerDL.R
\name{raw_flankerDL}
\alias{raw_flankerDL}
\title{Creates a "tidy" raw dataframe for the FlankerDL task}
\usage{
raw_flankerDL(x)
}
\arguments{
\item{x}{dataframe (an imported .emrge file)}
}
\description{
Creates a "tidy" raw dataframe for the FlankerDL task
}
