% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf.R
\name{raw_numberseries}
\alias{raw_numberseries}
\title{A Task Scoring Function}
\usage{
raw_numberseries(x)
}
\arguments{
\item{x}{a .txt exported E-Merge (or single E-Data) file}
}
\description{
Creates a raw data file of the NumberSeries task from the E-Merge file
}
\examples{
raw_numberseries(data)
}
\keyword{raw}
