% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospan.R
\name{raw_ospan}
\alias{raw_ospan}
\title{Creates a "tidy" raw dataframe for the Operation Span task}
\usage{
raw_ospan(x, blocks = NULL, taskVersion = "new", keep_col = c())
}
\arguments{
\item{x}{dataframe}

\item{blocks}{depricated. No need to indicate number of blocks}

\item{taskVersion}{old or new version. Required for different types of
data files. First try leaving out the argument, if it does not work
then try taskVersion = "old".}

\item{keep_col}{List of extra columns to keep}
}
\description{
Creates a "tidy" raw dataframe for the Operation Span task
}
