% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_rapm.R
\name{raw_rapm}
\alias{raw_rapm}
\title{A Task Scoring Function}
\usage{
raw_rapm(x)
}
\arguments{
\item{x}{a .txt exported E-Merge (or single E-Data) file}
}
\description{
Creates a raw data file of the RAPM task from the E-Merge file
}
\examples{
raw_rapm(data)
}
\keyword{raw}
