% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotspan.R
\name{raw_rotspan}
\alias{raw_rotspan}
\title{Raw Tidy Data for Rotation Span}
\usage{
raw_rotspan(x, include_col = c(), taskVersion = "new")
}
\arguments{
\item{x}{dataframe}

\item{include_col}{c(): list of additional columns to include}

\item{taskVersion}{old or new version. Required for different types of
data files from older or newer versions of the complex-span tasks.
First leave out the argument, if it does not work
then try taskVersion = "old".}
}
\description{
Converts the messy e-prime data file into a tidy raw data file that is
easy to work with.
}
