% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotspan.R
\name{raw_rotspan}
\alias{raw_rotspan}
\title{Creates a "tidy" raw dataframe for the Rotation Span task}
\usage{
raw_rotspan(x, blocks = NULL, taskVersion = "new", keep_col = c())
}
\arguments{
\item{x}{dataframe}

\item{blocks}{depricated. No need to indicate number of blocks}

\item{taskVersion}{old or new version. Required for different types of
data files. First try leaving out the argument, if it does not work
then try taskVersion = "old".}

\item{keep_col}{List of extra columns to keep}
}
\description{
Creates a "tidy" raw dataframe for the Rotation Span task
}
