% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sact.R
\name{raw_sact}
\alias{raw_sact}
\title{Creates a "tidy" raw dataframe for the SACT task}
\usage{
raw_sact(x)
}
\arguments{
\item{x}{dataframe (an imported .emrge file)}
}
\description{
Creates a "tidy" raw dataframe for the SACT task
}
