% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stroopDL.R
\name{raw_stroopDL}
\alias{raw_stroopDL}
\title{Raw Tidy Data for StroopDL}
\usage{
raw_stroopDL(x, include_col = c())
}
\arguments{
\item{x}{dataframe}

\item{include_col}{c(): list of additional columns to include}
}
\description{
Converts the messy e-prime data file into a tidy raw data file that is
easy to work with.
}
