% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symspan.R
\name{raw_symspan}
\alias{raw_symspan}
\title{Creates a "tidy" raw dataframe for the SymSpan task}
\usage{
raw_symspan(x, blocks = "", taskVersion = "new")
}
\arguments{
\item{x}{dataframe (an imported .emrge file)}

\item{blocks}{number of blocks administered. From 1-3}

\item{taskVersion}{old or new version. Old version means the
Procedure[Block] variable has a different label. (Default = "new")}
}
\description{
Creates a "tidy" raw dataframe for the SymSpan task
}
