% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_symspan.R
\name{raw_symspan}
\alias{raw_symspan}
\title{A Task Scoring Function}
\usage{
raw_symspan(x, blocks = "", taskVersion = "new")
}
\arguments{
\item{x}{a .txt exported E-Merge (or single E-Data) file}

\item{blocks}{number of blocks administered. From 1-3}

\item{taskVersion}{old or new version. Old version means the Procedure[Block] variable has a different label. (Default = "new")}
}
\description{
Creates a raw data file of the SymSpan task from an E-Merged file
}
\examples{
raw_symspan(data, blocks = 2)
}
\keyword{raw}
