% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualarrays.R
\name{raw_visualarrays}
\alias{raw_visualarrays}
\title{Raw Tidy Data for Visual Arrays}
\usage{
raw_visualarrays(x, include_col = c())
}
\arguments{
\item{x}{dataframe}

\item{include_col}{list of additional columns to include}
}
\description{
Converts the messy e-prime data file into a tidy raw data file that is
easy to work with.
}
