% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_outliers.R
\name{replace_outliers}
\alias{replace_outliers}
\title{Replace Outlier Scores}
\usage{
replace_outliers(x, variables = c(), cutoff = 3.5, with = "NA", id = "Subject")
}
\arguments{
\item{x}{dataframe}

\item{variables}{variables c() to be trimmed}

\item{cutoff}{zscore cutoff to use for trimming (default: 3.5)}

\item{with}{What value should the trimmed values be replaced with.
(default: replace = "NA")}

\item{id}{If variables = "all", then need to supply the subject ID variable}
}
\description{
This function will replace outliers variable(s) based on a zscore cutoff
}
