% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_outliers.R
\name{replace_outliers}
\alias{replace_outliers}
\title{Replace Outlier Scores}
\usage{
replace_outliers(
  x,
  variables = c(),
  cutoff = 3.5,
  with = "NA",
  id = "Subject",
  log_file = NULL
)
}
\arguments{
\item{x}{dataframe}

\item{variables}{variables c() to be trimmed}

\item{cutoff}{z-score cutoff to use for trimming (default: 3.5)}

\item{with}{What value should the trimmed values be replaced with.
(default: with = "NA")}

\item{id}{If variables = "all", then need to supply the subject ID variable}

\item{log_file}{the file path and name for where to save a log file to}
}
\description{
This function will replace outliers based on a z-score cutoff
}
