% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_problematic.R, R/replace_outliers.R
\name{replace_outliers}
\alias{replace_outliers}
\title{Remove Problematic Subjects}
\usage{
replace_outliers(
  x,
  variables = c(),
  cutoff = 3.5,
  with = "NA",
  id = "Subject",
  log_file = NULL
)

replace_outliers(
  x,
  variables = c(),
  cutoff = 3.5,
  with = "NA",
  id = "Subject",
  log_file = NULL
)
}
\arguments{
\item{x}{dataframe}

\item{variables}{variables c() to be trimmed}

\item{cutoff}{zscore cutoff to use for trimming (default: 3.5)}

\item{with}{What value should the trimmed values be replaced with.
(default: replace = "NA")}

\item{id}{If variables = "all", then need to supply the subject ID variable}

\item{log_file}{the file path and name for where to save a log file to}

\item{filter}{an expression passed on to dplyr:filter()}
}
\description{
This function will remove problematic subjects based on a filter criteria
and save them to a log data file.

This function will replace outliers variable(s) based on a zscore cutoff
}
