% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf.R
\name{score_lettersets}
\alias{score_lettersets}
\title{A Task Scoring Function}
\usage{
score_lettersets(x)
}
\arguments{
\item{x}{a .txt exported E-Merge (or single E-Data) file}
}
\description{
Scores the Letter Sets task taking a .txt exported E-Merge (or single E-Data) file as input
}
\examples{
score_lettersets(data)
}
\keyword{score}
