% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf.R
\name{score_numberseries}
\alias{score_numberseries}
\title{Calculate NumberSeries scores from a messy raw dataframe}
\usage{
score_numberseries(x)
}
\arguments{
\item{x}{dataframe (an imported .emrge file)}
}
\description{
This function skips the 'raw_numberseries()' step and therefore
    is not advised. However, some researchers may find
    it easier to just skip right to 'score_numberseries()'
}
