% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospan.R
\name{score_ospan}
\alias{score_ospan}
\title{Calculate Operation Span scores from a messy raw dataframe}
\usage{
score_ospan(x, blocks = "")
}
\arguments{
\item{x}{dataframe}

\item{blocks}{depricated. No need to indicate number of blocks.
Use group_by(Subject, Block) instead}
}
\description{
Calculate Operation Span scores from a messy raw dataframe
}
