% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospan.R
\name{score_ospan}
\alias{score_ospan}
\title{Calculate Operation Span scores from a messy raw dataframe}
\usage{
score_ospan(x, blocks = "", keep_col = c())
}
\arguments{
\item{x}{dataframe}

\item{blocks}{depricated. No need to indicate number of blocks.
Use group_by(Subject, Block) instead}

\item{keep_col}{List of extra columns to keep}
}
\description{
This function skips the 'raw_ospan()' step and therefore
    is not advised. However, some researchers may find
    it easier to just skip right to 'score_ospan()'
}
