% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotspan.R
\name{score_rotspan}
\alias{score_rotspan}
\title{Calculate Rotation Span scores from a messy raw dataframe}
\usage{
score_rotspan(x, blocks = "", keep_col = c())
}
\arguments{
\item{x}{dataframe (an imported .emrge file)}

\item{blocks}{number of blocks administered. From 1-3}

\item{keep_col}{List of column names to keep in scored dataframe}
}
\description{
This function skips the 'raw_rotspan()' step and therefore
    is not advised. However, some researchers may find
    it easier to just skip right to 'score_rotspan()'
}
