% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotspan.R
\name{score_rotspan}
\alias{score_rotspan}
\title{Calculate Rotation Span Scores}
\usage{
score_rotspan(x)
}
\arguments{
\item{x}{dataframe}
}
\description{
Calculate various span scores from the output of `raw_rotspan()`
}
