% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_rotspan.R
\name{score_rotspan}
\alias{score_rotspan}
\title{A Task Scoring Function}
\usage{
score_rotspan(x, blocks = "")
}
\arguments{
\item{x}{a .txt exported E-Merge (or single E-Data) file}

\item{blocks}{number of blocks administered. From 1-3}
}
\description{
Scores the Rotation Span (RotSPAN) task taking a .txt exported E-Merge (or single E-Data) file as input
}
\examples{
score_rotspan(x, blocks = 2)
}
\keyword{score}
