% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stroopDL.R
\name{score_stroopDL}
\alias{score_stroopDL}
\title{Calculate StroopDL threshold scores from a messy raw dataframe}
\usage{
score_stroopDL(x)
}
\arguments{
\item{x}{dataframe (an imported .emrge file)}
}
\description{
This function skips the 'raw_stroopDL()' step and therefore
    is not advised. However, some researchers may find
    it easier to just skip right to 'score_stroopDL()'
}
