% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symspan.R
\name{score_symspan}
\alias{score_symspan}
\title{Calculate Symmetry Span Scores}
\usage{
score_symspan(x)
}
\arguments{
\item{x}{dataframe}
}
\description{
Calculate various span scores from the output of `raw_symspan()`
}
