% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_symspan.R
\name{score_symspan}
\alias{score_symspan}
\title{A Task Scoring Function}
\usage{
score_symspan(x, blocks = "")
}
\arguments{
\item{x}{a .txt exported E-Merge (or single E-Data) file}

\item{blocks}{number of blocks administered. From 1-3}
}
\description{
Scores the Symmetry Span (SSPAN) task taking a .txt exported E-Merge (or single E-Data) file as input
}
\examples{
score_symspan(data, blocks = 2)
}
\keyword{score}
