% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symspan.R
\name{score_symspan}
\alias{score_symspan}
\title{Calculate Symmetry Span scores from a messy raw dataframe}
\usage{
score_symspan(x, blocks = NULL)
}
\arguments{
\item{x}{dataframe}

\item{blocks}{depricated. No need to indicate number of blocks.
Use group_by(Subject, Block) instead}
}
\description{
This function skips the 'raw_symspan()' step and therefore
    is not advised. However, some researchers may find
    it easier to just skip right to 'score_symspan()'
}
