% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualarrays.R
\name{score_visualarrays}
\alias{score_visualarrays}
\title{Calculate Visual Array k scores from a messy raw dataframe}
\usage{
score_visualarrays(x)
}
\arguments{
\item{x}{dataframe (an imported .emrge file)}
}
\description{
This function skips the 'raw_visualarrays()' step and therefore
    is not advised. However, some researchers may find
    it easier to just skip right to 'score_visualarrays()'
}
