% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualarrays.R
\name{score_visualarrays}
\alias{score_visualarrays}
\title{Calculate Visual Array k scores from a messy raw dataframe}
\usage{
score_visualarrays(x, taskname = "VAorient_S", id_col = "Subject")
}
\arguments{
\item{x}{dataframe (an imported .emrge file)}

\item{taskname}{string to add as a prefix to columns}

\item{id_col}{Subject id column}
}
\description{
This function will calculate a k score from the raw data
outputed by raw_visualarrays().
}
